implement horizontal flipping in renderers
