+++ packages/jbrowse-web/src/plugins/LinearGenomeView/models/configSchema.js
@@ -3,14 +3,15 @@ import { ConfigurationSchema } from '../../../configuration'
 export default ConfigurationSchema(
   'LinearGenomeView',
   {
+    reversed: {
+      type: 'boolean',
+      defaultValue: false,
+      description: 'horizontally flip the view',
+    },
     trackSelectorType: {
       type: 'string',
       defaultValue: 'hierarchical',
     },
-    // defaultDisplayRegions: {
-    //   type: 'frozenArray',
-    //   defaultValue: [],
-    // },
   },
   { singleton: true },
 )
